fix (Audit && DeleteEnterprise): serializer will acept ImageFile type
After manual tests, I found out that were a bug in
ExcludeEnterpriseView. The bug was not in his logic, but on the Audit
signal linked to it.

I've made a function before the attribute changes_audit were made, so
the system checks if there is any data of types: FileField, ImageField,
Decimal and Datetime. If any, the check func convert to its name, real
type or format - in case of datetime.