feat (GoogleLoginAuth): Login by google account.
This feature links the ArquiVia account to the user Google account, if
the email is already in use. Otherwise, the user access it ArquiVia
account created using google (no diff).

If there's no such email on database, a new account will be created.
(Account creation using Google still not developed and tested).